refactor: mutations for job
