+++ packages/shared/src/components/opportunity/OpportunityEditModal/OpportunityEditContentModal.tsx
@@ -11,7 +11,7 @@ import { Loader } from '../../Loader';
 
 import { Button, ButtonSize, ButtonVariant } from '../../buttons/Button';
 import { labels } from '../../../lib';
-import { editOpportunityContentMutationOptions } from '../../../features/opportunity/graphql';
+import { editOpportunityContentMutationOptions } from '../../../features/opportunity/mutations';
 import { ApiError } from '../../../graphql/common';
 import { useUpdateQuery } from '../../../hooks/useUpdateQuery';
 import { useToastNotification } from '../../../hooks';
